api/training/requirements.txt
attrs==18.2.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click-completion==0.5.2
click-didyoumean==0.0.3
click-help-colors==0.8
colorama==0.3.9
convertdate==2.1.3
cryptography==2.8
cursor==1.3.4
cycler==0.10.0
Cython==0.29.15
ephem==3.7.7.1
flask==1.1.1
gradient==0.6.0
halo==0.0.29
holidays==0.10.1
idna==2.9
Jinja2==2.11.1
kiwisolver==1.1.0
log-symbols==0.0.14
LunarCalendar==0.0.9
MarkupSafe==1.1.1
marshmallow==2.21.0
matplotlib==3.2.0
numpy==1.18.1
oauthlib==3.1.0
pandas==1.0.2
progressbar2==3.50.0
prompt-toolkit==2.0.10
pycparser==2.20
pyOpenSSL==19.1.0
pyparsing==2.4.6
pystan==2.19.0.0
python-dateutil==2.8.1
python-dotenv==0.12.0
python-utils==2.4.0
pytz==2019.3
PyYAML==5.3
redis
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
requests==2.23.0
shellingham==1.3.2
six==1.14.0
spinners==0.0.24
termcolor==1.1.0
terminaltables==3.1.0
tornado==6.0.4
urllib3==1.25.8
wcwidth==0.1.8
scikit-learn==0.22.1